# Faster parquet utf8 validation using `simdjson`