--- a/parquet/Cargo.toml
+++ b/parquet/Cargo.toml
@@ -69,6 +69,7 @@ paste = { version = "1.0" }
 half = { version = "2.1", default-features = false, features = ["num-traits"] }
 sysinfo = { version = "0.32.0", optional = true, default-features = false, features = ["system"] }
 crc32fast = { version = "1.4.2", optional = true, default-features = false }
+simdutf8 = { workspace = true }
 
 [dev-dependencies]
 base64 = { version = "0.22", default-features = false, features = ["std"] }