@tustvold 2024-10-31
I'm not sure that 5% really justifies an additional dependency, especially one that uses so much unsafe...