@Dandandan 2024-11-01
FWIW some other projects are using `simdutf8` as well, like polars https://github.com/pola-rs/polars/blob/main/Cargo.toml#L77