@alamb 2024-11-07
I am not sure exactly the usecase here, but what about simply disabling utf8 validation for known good data?
- https://github.com/apache/arrow-rs/issues/6701