@Dandandan 2024-11-08
> I am not sure exactly the usecase here, but what about simply disabling utf8 validation for known good data?
> 
> * [Proposal: Add `unsafe` option to disable UTF8 validation on parquet read #6701](https://github.com/apache/arrow-rs/issues/6701)

The "use case" of this PR is just that utf8 validation takes time, this PR improves the performance.