@alamb 2024-12-17
So what shall we do with this PR? Make it an optional opt-in feature of the parquet crate that people can enable if they want more performance?