@doki23 2024-12-24
I believe this PR is solely for performance enhancement. Introducing an optional opt-in feature deserves a separate PR.