@XiangpengHao 2025-01-01
> Improves performance for about 4-5% (on M1 Pro) on strings (plain encoding)

Coming from https://github.com/apache/arrow-rs/pull/6921#discussion_r1900411077, I have seen much larger performance (~15%) improvements using simdutf8 with StringViewArray + x86, especially when strings are long (>128 byte).