@XiangpengHao 2025-01-02
This is my benchmark results with Clickbench non-partitioned and filter pushdown enabled. Benchmarked on x86 AMD 9900x. Scan-dominate queries can get 20% improvements.

```
┏━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━┳━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃ no-simd-utf8 ┃ simd-utf8 ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━╇━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 0     │       0.35ms │    0.36ms │     no change │
│ QQuery 1     │      50.50ms │   50.15ms │     no change │
│ QQuery 2     │      68.05ms │   67.37ms │     no change │
│ QQuery 3     │      89.46ms │   88.81ms │     no change │
│ QQuery 4     │     463.69ms │  459.30ms │     no change │
│ QQuery 5     │     527.10ms │  481.84ms │ +1.09x faster │
│ QQuery 6     │      50.25ms │   50.92ms │     no change │
│ QQuery 7     │      55.59ms │   55.34ms │     no change │
│ QQuery 8     │     562.75ms │  561.54ms │     no change │
│ QQuery 9     │     580.23ms │  580.26ms │     no change │
│ QQuery 10    │     159.95ms │  155.25ms │     no change │
│ QQuery 11    │     172.33ms │  168.08ms │     no change │
│ QQuery 12    │     730.38ms │  653.11ms │ +1.12x faster │
│ QQuery 13    │    1205.87ms │ 1100.43ms │ +1.10x faster │
│ QQuery 14    │     739.05ms │  666.65ms │ +1.11x faster │
│ QQuery 15    │     550.73ms │  552.73ms │     no change │
│ QQuery 16    │    1188.05ms │ 1159.64ms │     no change │
│ QQuery 17    │    1159.66ms │ 1146.29ms │     no change │
│ QQuery 18    │    2641.59ms │ 2652.32ms │     no change │
│ QQuery 19    │      81.54ms │   90.09ms │  1.10x slower │
│ QQuery 20    │     610.90ms │  589.75ms │     no change │
│ QQuery 21    │     705.48ms │  663.11ms │ +1.06x faster │
│ QQuery 22    │    1659.48ms │ 1357.19ms │ +1.22x faster │
│ QQuery 23    │    3534.78ms │ 3639.67ms │     no change │
│ QQuery 24    │     533.81ms │  475.18ms │ +1.12x faster │
│ QQuery 25    │     470.34ms │  373.89ms │ +1.26x faster │
│ QQuery 26    │     576.26ms │  476.26ms │ +1.21x faster │
│ QQuery 27    │    1121.43ms │ 1056.52ms │ +1.06x faster │
│ QQuery 28    │    4291.35ms │ 4288.26ms │     no change │
│ QQuery 29    │     228.93ms │  235.60ms │     no change │
│ QQuery 30    │     589.98ms │  541.70ms │ +1.09x faster │
│ QQuery 31    │     716.59ms │  702.51ms │     no change │
│ QQuery 32    │    2593.06ms │ 2528.18ms │     no change │
│ QQuery 33    │    2362.52ms │ 2336.75ms │     no change │
│ QQuery 34    │    2360.44ms │ 2334.17ms │     no change │
│ QQuery 35    │     705.78ms │  696.32ms │     no change │
│ QQuery 36    │     163.69ms │  156.20ms │     no change │
│ QQuery 37    │     129.14ms │   97.25ms │ +1.33x faster │
│ QQuery 38    │      88.72ms │   90.76ms │     no change │
│ QQuery 39    │     218.61ms │  213.61ms │     no change │
│ QQuery 40    │      71.73ms │   73.40ms │     no change │
│ QQuery 41    │      68.23ms │   67.56ms │     no change │
│ QQuery 42    │      69.22ms │   68.76ms │     no change │
└──────────────┴──────────────┴───────────┴───────────────┘
┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━┓
┃ Benchmark Summary           ┃            ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━┩
│ Total Time (no-simd-utf8)   │ 34947.62ms │
│ Total Time (simd-utf8)      │ 33803.08ms │
│ Average Time (no-simd-utf8) │   812.74ms │
│ Average Time (simd-utf8)    │   786.12ms │
│ Queries Faster              │         12 │
│ Queries Slower              │          1 │
│ Queries with No Change      │         30 │
└─────────────────────────────┴────────────┘
```