@etseidl 2025-01-08
> @etseidl I wonder if you have any thoughts on this PR?

None that haven't already been voiced. It seems like a fairly low risk (especially if made optional) way to get a significant speed up in string handling.

+1