@alamb 2025-01-14
I merged this branch up from main and intend to merge it when the CI passes.

I will also make a follow on PR to make this feature optional.