@alamb 2025-01-14
Let's keep chipping away at performance 🚀  -- thank you all for your thoughts and help here

Here is the follow on PR to add validation as a feature flag:
- https://github.com/apache/arrow-rs/pull/6979